pyproject.toml
[project]
name = "notion-gcal-slack"
version = "0.1.0"
description = "Notion Todo to Google Calendar sync worker"
authors = [
    {name = "Lukael",email = "dev@lukael.kr"}
]
readme = "README.md"
requires-python = ">=3.11, <3.14"
dependencies = [
    "python-dotenv (>=1.2.2,<2.0.0)",
    "notion-client (>=3.0.0,<4.0.0)",
    "google-api-python-client (>=2.192.0,<3.0.0)",
    "google-auth-oauthlib (>=1.3.0,<2.0.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
black = ">=26.3.0,<27.0.0"
pylint = ">=4.0.5,<5.0.0"
pytest = ">=8.4.2,<9.0.0"
pytest-cov = ">=7.0.0,<8.0.0"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.pylint.main]
ignore-paths = [
  "^tests/.*$"
]

[tool.pylint.format]
max-line-length = 100

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=85"
testpaths = ["tests"]
